Fixed deprecated network reachability function warnings.

+++ portmap/AppController.m
@@ -163,7 +163,7 @@
     [self updateTagLine];
 }
 
-- (void)genericSheetDidEnd:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo {
+- (void)genericSheetDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo {
     [sheet orderOut:self];
 }
